Separate ConfigureSearchTable into two components
