+++ context/app/static/js/components/entity-search/sidebar/ConfigureSearchTableWrapper/index.js
@@ -1,0 +1,3 @@
+import ConfigureSearchTableWrapper from './ConfigureSearchTableWrapper';
+
+export default ConfigureSearchTableWrapper;
