+++ context/app/static/js/components/entity-search/sidebar/ConfigureSearchTableWrapper/style.js
