+++ context/app/static/js/components/entity-search/sidebar/ConfigureSearchTableWrapper/utils.js
