+++ context/app/static/js/components/entity-search/sidebar/DonorSampleConfigureSearchTable/DonorSampleConfigureSearchTable.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 
-import ConfigureSearchTable from 'js/components/entity-search/sidebar/ConfigureSearchTable';
+import ConfigureSearchTableWrapper from 'js/components/entity-search/sidebar/ConfigureSearchTableWrapper';
 import { useDonorSampleMetadatFieldConfigs } from './hooks';
 
 function DonorSampleConfigureSearchTable(props) {
   const metadataFields = useDonorSampleMetadatFieldConfigs();
 
-  return <ConfigureSearchTable {...props} availableFields={metadataFields} />;
+  return <ConfigureSearchTableWrapper {...props} availableFields={metadataFields} />;
 }
 
 export default DonorSampleConfigureSearchTable;
